@@ -1,19 +1,17 @@
 ﻿using Microsoft.Extensions.Diagnostics.HealthChecks;
-using System.Threading;
-using System.Threading.Tasks;
+
 using WebSampleApp.Services;
 
-namespace WebSampleApp
-{
-    public class CustomHealthCheck : IHealthCheck
-    {
-        private readonly HealthSample _healthSample;
-        public CustomHealthCheck(HealthSample healthSample) => _healthSample = healthSample;
+namespace WebSampleApp;
 
-        public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
-        {
-            if (_healthSample.IsHealthy) return Task.FromResult(HealthCheckResult.Healthy("healthy"));
-            else return Task.FromResult(HealthCheckResult.Unhealthy("unhealthy"));
-        }
+public class CustomHealthCheck : IHealthCheck
+{
+    private readonly HealthSample _healthSample;
+    public CustomHealthCheck(HealthSample healthSample) => _healthSample = healthSample;
+
+    public Task<HealthCheckResult> CheckHealthAsync(HealthCheckContext context, CancellationToken cancellationToken = default)
+    {
+        if (_healthSample.IsHealthy) return Task.FromResult(HealthCheckResult.Healthy("healthy"));
+        else return Task.FromResult(HealthCheckResult.Unhealthy("unhealthy"));
     }
 }
