@@ -1,17 +1,12 @@
 ﻿using Microsoft.AspNetCore.Components;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace WebSampleApp
+namespace WebSampleApp;
+
+[Route("Foo")]
+public class MyController
 {
-    [Route("Foo")]
-    public class MyController
+    public MyController()
     {
-        public MyController()
-        {
 
-        }
     }
 }
